+++ styles/globals.css
@@ -256,3 +256,7 @@ text-xs transition-all group-hover:opacity-60;
 .search-box-icon {
   @apply h-5 w-5
 }
+
+.error-message {
+  @apply pt-0 text-xs leading-8 text-red-500 md:pt-1 md:text-xs md:leading-tight lg:w-[200px];
+}
